requirements.txt
python-telegram-bot==20.5
requests==2.26.0
